requirements.txt
torch
torchaudio
pyaudio
wave
deepfilternet
pyannote.audio
numpy
